% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bradford_broadband_1820.R
\docType{data}
\name{bradford_broadband_1820}
\alias{bradford_broadband_1820}
\title{Prices of 50,000 round cut diamonds.}
\format{
A data frame with 95,258 rows and 26 variables:
\describe{
\item{postcode}{postcode, Bradford postcode}
\item{pc1}{postcode's 1st part, Bradford postcode}
\item{pc2}{postcode's 2nd part, Bradford postcode}
\item{year}{year of survey}
\item{month}{month of survey}
...
}
}
\source{
\url{https://www.ofcom.org.uk/research-and-data}
}
\usage{
bradford_broadband_1820
}
\description{
A dataset containing the Ofcom's broadband survey data for Bradford city
between 2018 and 2020.
}
\keyword{datasets}
