% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imd.R
\docType{data}
\name{imd}
\alias{imd}
\title{Index of Multiple Deprivation in Bradford}
\format{
A data frame with 417 rows and 43 variables:
\describe{
\item{LSOA11CD}{LSOA code}
\item{LSOA11NM}{LSOA name}
\item{LAD19CD}{Local Authority Districts code}
\item{LAD19NM}{Local Authority Districts name}
\item{IMD19}{Index of Multiple Deprivation 2019}
\item{FID}{id}

}
}
\source{
\url{https://www.gov.uk/government/collections/english-indices-of-deprivation}
}
\usage{
imd
}
\description{
A dataset containing Index of Multiple Deprivation 2019 for Bradford city
}
\keyword{datasets}
