% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/property.R
\name{property}
\alias{property}
\title{Load property price paid data}
\source{
\url{https://www.gov.uk/government/statistical-data-sets/price-paid-data-downloads}
}
\usage{
property(var1)
}
\arguments{
\item{var1}{empty}
}
\value{
character
}
\description{
Load property price paid data
}
\examples{
property()
}
