% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_school_admission.R
\docType{data}
\name{sec_school_admission}
\alias{sec_school_admission}
\title{Secondary school admission in Bradford}
\format{
A data frame with 417 rows and 43 variables:
\describe{
\item{id}{id}

}
}
\source{
\url{https://www.gov.uk/school-performance-tables}
}
\usage{
sec_school_admission
}
\description{
A dataset containing secondary school admission data for Bradford city
between 2014 and 2020.
}
\keyword{datasets}
