% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_school_census.R
\docType{data}
\name{sec_school_census}
\alias{sec_school_census}
\title{Secondary school census in Bradford}
\format{
A data frame with 417 rows and 36 variables:
\describe{
\item{id}{id}

}
}
\source{
\url{https://www.gov.uk/school-performance-tables}
}
\usage{
sec_school_census
}
\description{
A dataset containing secondary school census data for Bradford city
between 2014 and 2020.
}
\keyword{datasets}
