% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_school_profile.R
\docType{data}
\name{sec_school_profile}
\alias{sec_school_profile}
\title{Secondary school profile in Bradford}
\format{
A data frame with 417 rows and 16 variables:
\describe{
\item{id}{id}

}
}
\source{
\url{https://www.gov.uk/school-performance-tables}
}
\usage{
sec_school_profile
}
\description{
A dataset containing secondary school profile data for Bradford city
between 2014 and 2020.
}
\keyword{datasets}
