% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wardmap.R
\docType{data}
\name{wardmap}
\alias{wardmap}
\title{Ward level shape file for Bradford}
\format{
A data frame with 30,000 rows and 10 variables:
\describe{
\item{objectid}{objectid}
\item{wd20cd}{Ward code}
\item{wd20nm}{Ward name}
\item{wd20nmw}{Ward NA}
\item{bng_e}{Easting}
\item{bng_n}{Northing}
\item{long}{longitude}
\item{lat}{Latitude}
\item{shape_are}{Shape area}
\item{shape_len}{Shape length}
\item{geometry}{geometry data}

}
}
\source{
\url{https://geoportal.statistics.gov.uk/datasets/wards-december-2020-uk-bfc-v2}
}
\usage{
wardmap
}
\description{
A dataset containing the ward level geometry for Bradford
}
\keyword{datasets}
